split file up into smaller modules
