+++ README.md
@@ -44,3 +44,5 @@ You can also change the bitrate of the output file by tweaking `-b` (Default is 
     - next, the width, then the height at that keyframe (in pixels)
     - finally, the interpolation with which to advance towards the next keyframe (currently, only `linear` is supported.)
   - If it isn't overwritten, an implicit keyframe at 0 frames into the video is added with linear interpolation and the video's original size.
+
+Additionally, any 2 modes can be combined using a `+` symbol, like `Bounce+Shutter`. If one only specifies width, and one only specifies height, then those respective values are used. If there is a conflict, the value from the latter of the 2 modes is used (so `Shrink+Bounce` would result in the same effect as just `Bounce`)
