+++ modes/shrink.js
@@ -1,0 +1,6 @@
+module.exports = {
+	setup: () => {},
+	getFrameBounds: (info) => ({
+		height: Math.max(1, Math.floor(info.maxHeight - (info.frame / info.frameCount) * info.maxHeight)),
+	}),
+}
